moved geofence to database package
