+++ streampipes-processors-geo-jvm/src/main/java/org/streampipes/processors/geo/jvm/database/geofence/storing/StoreGeofenceController.java
@@ -1,4 +1,4 @@
-package org.streampipes.processors.geo.jvm.geofence.storing;
+package org.streampipes.processors.geo.jvm.database.geofence.storing;
 
 import org.streampipes.model.DataSinkType;
 import org.streampipes.model.graph.DataSinkDescription;
@@ -28,7 +28,7 @@ public class StoreGeofenceController extends StandaloneEventSinkDeclarer<StoreGe
 
   @Override
   public DataSinkDescription declareModel() {
-    return DataSinkBuilder.create("org.streampipes.processors.geo.jvm.geofence", EPA_NAME,
+    return DataSinkBuilder.create("org.streampipes.processors.geo.jvm.database.geofence.storing", EPA_NAME,
         "Stores Polygons or MultiPolygons as geofence together with calculated area in a database.")
             .category(DataSinkType.STORAGE)
             .withAssets(Assets.DOCUMENTATION, Assets.ICON)
