+++ streampipes-processors-geo-jvm/src/main/resources/org.streampipes.processors.geo.jvm.database.geofence.enricher/icon.png
