+++ streampipes-processors-geo-jvm/src/main/resources/org.streampipes.processors.geo.jvm.database.geofence.storing/strings.en
